chore: Setup routes and module boilerplate
