+++ api.planx.uk/modules/flows/service/moveFlow.test.ts
@@ -1,8 +1,8 @@
 import supertest from "supertest";
 
-import { queryMock } from "../tests/graphqlQueryMock";
-import { authHeader } from "../tests/mockJWT";
-import app from "../server";
+import { queryMock } from "../../../tests/graphqlQueryMock";
+import { authHeader } from "../../../tests/mockJWT";
+import app from "../../../server";
 
 beforeEach(() => {
   queryMock.mockQuery({
